requirements.txt
fastapi
uvicorn
anthropic
httpx
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
